requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
alembic==1.15.2
altgraph==0.17.4
anyio==4.8.0
attrs==25.1.0
azure-core==1.32.0
azure-identity==1.20.0
azure-storage-blob==12.25.1
babel==2.17.0
bcrypt==4.1.2
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cryptography==44.0.1
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
Deprecated==1.2.18
et_xmlfile==2.0.0
Flask==3.0.2
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
fonttools==4.56.0
frozenlist==1.5.0
future==1.0.0
geocoder==1.38.1
greenlet==3.1.1
gunicorn==21.2.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.0.0
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.1
microsoft-kiota-abstractions==1.9.2
microsoft-kiota-authentication-azure==1.9.2
microsoft-kiota-http==1.9.2
microsoft-kiota-serialization-form==1.9.2
microsoft-kiota-serialization-json==1.9.2
microsoft-kiota-serialization-multipart==1.9.2
microsoft-kiota-serialization-text==1.9.2
msal==1.31.1
msal-extensions==1.2.0
msgraph-core==1.3.1
multidict==6.1.0
numpy==1.26.4
openpyxl==3.1.5
opentelemetry-api==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
packaging==24.2
pandas==2.2.1
pefile==2023.2.7
pillow==11.1.0
pluggy==1.5.0
portalocker==2.10.1
propcache==0.2.1
psutil==6.1.1
psycopg2-binary==2.9.10
pycparser==2.22
pycryptodome==3.21.0
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
PyJWT==2.10.1
pyparsing==3.2.3
pytest==8.3.4
python-dateutil==2.9.0.post0
pytz==2024.2
qrcode==8.0
ratelim==0.1.6
reportlab==4.1.0
requests==2.32.3
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
std-uritemplate==2.0.3
tkcalendar==1.6.1
ttkbootstrap==1.10.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
yarl==1.18.3
zipp==3.21.0
